move api functions in one file and delete token class
