--- client/src/api/Token.js
@@ -1,9 +1,0 @@
-export class Token {
-  token;
-
-  constructor(token) {
-    this.token = token;
-    // TODO I can token's payload as fields of the class
-
-  }
-}
